+++ lib/win32audio.dart
@@ -16,7 +16,7 @@ class AudioDevice {
 
   //tomap
   Map<String, dynamic> toMap() {
-    return {
+    return <String, dynamic>{
       'id': id,
       'name': name,
       'iconPath': iconPath,
@@ -50,7 +50,7 @@ class ProcessVolume {
 
   // tomap
   Map<String, dynamic> toMap() {
-    return {
+    return <String, dynamic>{
       'processId': processId,
       'processPath': processPath,
       'maxVolume': maxVolume,
@@ -76,21 +76,21 @@ enum AudioDeviceType {
   input,
 }
 
-const audioMethodChannel = MethodChannel('win32audio');
+const MethodChannel audioMethodChannel = MethodChannel('win32audio');
 
 class Audio {
   /// Returns a Future list of audio devices of a specified type.
   /// The type is specified by the [AudioDeviceType] enum.
   ///
   static Future<List<AudioDevice>?> enumDevices(AudioDeviceType audioDeviceType) async {
-    final Map<String, dynamic> arguments = {'deviceType': audioDeviceType.index};
+    final Map<String, dynamic> arguments = <String, dynamic>{'deviceType': audioDeviceType.index};
     final Map<dynamic, dynamic> map = await audioMethodChannel.invokeMethod('enumAudioDevices', arguments);
-    List<AudioDevice>? audioDevices = [];
-    for (var key in map.keys) {
-      final audioDevice = AudioDevice();
+    List<AudioDevice>? audioDevices = <AudioDevice>[];
+    for (int key in map.keys) {
+      final AudioDevice audioDevice = AudioDevice();
       audioDevice.id = map[key]['id'];
       audioDevice.name = map[key]['name'];
-      final iconData = map[key]['iconInfo'].split(",");
+      final List<String> iconData = map[key]['iconInfo'].split(",");
       audioDevice.iconPath = iconData[0];
       audioDevice.iconID = int.parse(iconData[1]);
       audioDevice.isActive = map[key]['isActive'];
@@ -103,12 +103,12 @@ class Audio {
   /// The type is specified by the [AudioDeviceType] enum.
   ///
   static Future<AudioDevice?> getDefaultDevice(AudioDeviceType audioDeviceType) async {
-    final Map<String, dynamic> arguments = {'deviceType': audioDeviceType.index};
+    final Map<String, dynamic> arguments = <String, dynamic>{'deviceType': audioDeviceType.index};
     final Map<dynamic, dynamic> map = await audioMethodChannel.invokeMethod('getDefaultDevice', arguments);
-    final audioDevice = AudioDevice();
+    final AudioDevice audioDevice = AudioDevice();
     audioDevice.id = map['id'];
     audioDevice.name = map['name'];
-    final iconData = map['iconInfo'].split(",");
+    final List<String> iconData = map['iconInfo'].split(",");
     audioDevice.iconPath = iconData[0];
     audioDevice.iconID = int.parse(iconData[1]);
     audioDevice.isActive = map['isActive'];
@@ -117,17 +117,23 @@ class Audio {
 
   /// Sets the default audio device.
   /// The type is specified by the [AudioDeviceType] enum.
-  static Future<int> setDefaultDevice(String deviceID) async {
-    final Map<String, dynamic> arguments = {'deviceID': deviceID};
-    final result = await audioMethodChannel.invokeMethod<int>('setDefaultAudioDevice', arguments);
+  static Future<int> setDefaultDevice(String deviceID, {bool console = false, bool multimedia = true, bool communications = false}) async {
+    final Map<String, dynamic> arguments = <String, dynamic>{
+      'deviceID': deviceID,
+      'console': console,
+      'multimedia': multimedia,
+      'communications': communications,
+    };
+
+    final int? result = await audioMethodChannel.invokeMethod<int>('setDefaultAudioDevice', arguments);
     return result as int;
   }
 
   /// Returns the current volume for the given audio device type.
   /// The type is specified by the [AudioDeviceType] enum.
   static Future<double> getVolume(AudioDeviceType audioDeviceType) async {
-    final Map<String, dynamic> arguments = {'deviceType': audioDeviceType.index};
-    final result = await audioMethodChannel.invokeMethod<double>('getAudioVolume', arguments);
+    final Map<String, dynamic> arguments = <String, dynamic>{'deviceType': audioDeviceType.index};
+    final double? result = await audioMethodChannel.invokeMethod<double>('getAudioVolume', arguments);
     return result as double;
   }
 
@@ -136,24 +142,29 @@ class Audio {
   ///
   static Future<int> setVolume(double volume, AudioDeviceType audioDeviceType) async {
     if (volume > 1) volume = (volume / 100).toDouble();
-    final Map<String, dynamic> arguments = {'deviceType': audioDeviceType.index, 'volumeLevel': volume};
-    final result = await audioMethodChannel.invokeMethod<int>('setAudioVolume', arguments);
+    final Map<String, dynamic> arguments = <String, dynamic>{'deviceType': audioDeviceType.index, 'volumeLevel': volume};
+    final int? result = await audioMethodChannel.invokeMethod<int>('setAudioVolume', arguments);
     return result as int;
   }
 
   /// This function switches the audio device to the specified type. The type is specified by the [AudioDeviceType] enum.
-  static Future<double> switchDefaultDevice(AudioDeviceType audioDeviceType) async {
-    final Map<String, dynamic> arguments = {'deviceType': audioDeviceType.index};
-    final result = await audioMethodChannel.invokeMethod<double>('switchDefaultDevice', arguments);
-    return result as double;
+  static Future<bool> switchDefaultDevice(AudioDeviceType audioDeviceType, {bool console = false, bool multimedia = true, bool communications = false}) async {
+    final Map<String, dynamic> arguments = <String, dynamic>{
+      'deviceType': audioDeviceType.index,
+      'console': console,
+      'multimedia': multimedia,
+      'communications': communications,
+    };
+    final bool? result = await audioMethodChannel.invokeMethod<bool>('switchDefaultDevice', arguments);
+    return result as bool;
   }
 
   /// Returns a Future with a list of ProcessVolume objects containing information about all audio mixers.
   static Future<List<ProcessVolume>?> enumAudioMixer() async {
     final Map<dynamic, dynamic> map = await audioMethodChannel.invokeMethod('enumAudioMixer');
-    List<ProcessVolume>? processVolumes = [];
-    for (var key in map.keys) {
-      final processVolume = ProcessVolume();
+    List<ProcessVolume>? processVolumes = <ProcessVolume>[];
+    for (int key in map.keys) {
+      final ProcessVolume processVolume = ProcessVolume();
       processVolume.processId = key;
       processVolume.processPath = map[key]['processPath'];
       processVolume.maxVolume = map[key]['maxVolume'];
@@ -168,8 +179,8 @@ class Audio {
   /// The process ID can be obtained by calling the enumAudioMixer function.
   static Future<bool> setAudioMixerVolume(int processID, double volume) async {
     if (volume > 1) volume = (volume / 100).toDouble();
-    final Map<String, dynamic> arguments = {'processID': processID, 'volumeLevel': volume};
-    final result = await audioMethodChannel.invokeMethod<bool>('setAudioMixerVolume', arguments);
+    final Map<String, dynamic> arguments = <String, dynamic>{'processID': processID, 'volumeLevel': volume};
+    final bool? result = await audioMethodChannel.invokeMethod<bool>('setAudioMixerVolume', arguments);
     return result as bool;
   }
 }
@@ -178,9 +189,10 @@ class Audio {
 /// The icon location is the path to the icon file.
 /// The icon ID is the icon ID of the icon file.
 /// The icon ID can be obtained by calling the enumAudioDevices function.
+@Deprecated("Use WinIcons class instead")
 Future<Uint8List?> nativeIconToBytes(String iconlocation, {int iconID = 0}) async {
-  final Map<String, dynamic> arguments = {'iconLocation': iconlocation, 'iconID': iconID};
-  final result = await audioMethodChannel.invokeMethod<Uint8List>('iconToBytes', arguments);
+  final Map<String, dynamic> arguments = <String, dynamic>{'iconLocation': iconlocation, 'iconID': iconID};
+  final Uint8List? result = await audioMethodChannel.invokeMethod<Uint8List>('iconToBytes', arguments);
   return result;
 }
 
@@ -189,11 +201,12 @@ class WinIcons {
   /// This function extracts an icon from any file.
 
   Future<Uint8List?> extractFileIcon(String iconlocation, {int iconID = 0}) async {
-    final Map<String, dynamic> arguments = {'iconLocation': iconlocation, 'iconID': iconID};
-    final result = await audioMethodChannel.invokeMethod<Uint8List>('iconToBytes', arguments);
+    final Map<String, dynamic> arguments = <String, dynamic>{'iconLocation': iconlocation, 'iconID': iconID};
+    final Uint8List? result = await audioMethodChannel.invokeMethod<Uint8List>('iconToBytes', arguments);
     return result;
   }
 
+  @Deprecated("Use extractFileIcon instead")
   Future<Uint8List?> extractExecutableIcon(String iconlocation, {int iconID = 0}) async {
     return await extractFileIcon(iconlocation, iconID: iconID);
   }